Move Search Functions to Own Store
