+++ src/store/TranscriptionsStore.js
@@ -1,18 +1,6 @@
 import { flow, getRoot, types } from 'mobx-state-tree'
 import ASYNC_STATES from 'helpers/asyncStates'
 
-const IDS = {
-  ZOONIVERSE: 'ZOONIVERSE ID',
-  INTERNAL: 'INTERNAL ID'
-}
-
-const STATUS = {
-  unseen: 0,
-  ready: 1,
-  approved: 2,
-  in_progress: 3
-}
-
 const Transcription = types.model('Transcription', {
   id: types.identifier,
   flaggedid: types.optional(types.boolean, false),
@@ -76,56 +64,9 @@ const TranscriptionsStore = types.model('TranscriptionsStore', {
     yield self.retrieveTranscriptions(`/transcriptions?filter[group_id_eq]=${groupName}`)
   }),
 
-  fetchTranscriptionsByFilter: flow(function * fetchTranscriptionsByFilter(args, group) {
-    let query = ''
-    const approvalFilters = ['unseen', 'in_progress', 'ready', 'approved']
-    const additionalFilters = ['flagged', 'low_consensus']
-    const activeApprovalFilters = []
-    const activeAdditionalFilters = []
-
-    Object.keys(args).forEach(key => {
-      if (args[key]) {
-        if (approvalFilters.includes(key)) activeApprovalFilters.push(key)
-        if (additionalFilters.includes(key)) activeAdditionalFilters.push(key)
-      }
-    })
-
-    if (activeApprovalFilters.length > 0) {
-      query += 'filter[status_in]='
-      activeApprovalFilters.forEach(filter => query += `${STATUS[filter]},`)
-    }
-
-    if (activeAdditionalFilters.length > 0) {
-      if (query.length > 0) query += '&'
-      activeAdditionalFilters.forEach(filter => query += `filter[${filter}_eq]=true&`)
-    }
-
-    yield self.retrieveTranscriptions(`/transcriptions?filter[group_id_eq]=${group}&${query}`)
-    getRoot(self).modal.toggleModal('')
-  }),
-
-  fetchTranscriptionsById: flow(function * fetchTranscriptionsById(type, value, group) {
-    type = type === IDS.ZOONIVERSE ? 'subject_id' : 'internal_id'
-    yield self.retrieveTranscriptions(`/transcriptions?filter[${type}_eq]=${value}&filter[group_id_eq]=${group}`)
-    getRoot(self).modal.toggleModal('')
-  }),
-
   reset: () => {
     self.selectTranscription(null)
     self.all.clear()
-  },
-
-  searchTranscriptions: function searchTranscriptions(args) {
-    const group = getRoot(self).groups.title
-    const idType = args.type === IDS.ZOONIVERSE || args.type === IDS.INTERNAL
-    const idValue = args.id && args.id.length > 0
-    if (idType && idValue) {
-      self.reset()
-      self.fetchTranscriptionsById(args.type, args.id, group)
-    } else {
-      self.reset()
-      self.fetchTranscriptionsByFilter(args, group)
-    }
   },
 
   selectTranscription: flow(function * selectTranscription(id = null) {
